Make within() global in both browser and nodejs
Using var at the top level only creates a global variable in the browser,
not in node (the variable is scoped  in the file module).

Except in strict mode (when "use strict"; is found in the outer scope
of the function), "this" refers to the global object in a function called
directly or applied to a null object.

Changing
var within = privately(function(){
  ...
  return within;
});

to

privately(function(){
  ...
  this.within = within;
});

exports within to the global scope in both browser environment and
nodejs environment.