Update documentation and unit tests of subscribe() before updating behavior
The callback will no longer be triggered immediately when the current
value is null or undefined. The change of behavior in this edge case aims
to make it possible to publish an event for which subscribers will only get
notified of the next occurrence, not the current one.

For example:

publish("tick"); // sets tick as undefined

// not fired immediately
var cancel = subscribe("tick", function runOnce() {
  cancel();
  ... // triggered only once, the next time "tick" is published
});

// fires runOnce
publish("tick");