Describe three forms of within() called with 0, 1 or 2 parameters
The upcoming form with 0 parameter provides a more convenient approach for the
use case of separate spaces for different instances of an object collection.

I removed the method proposed previously, dynamic module names, and described
simpler alternatives, dynamic event names and "id" properties in event data
as alternatives to the new approach.