Complete documentation of within()
I updated the documentation to describe the anonymous modules.

I also included examples of the three forms of the API in the Usage
section at the top of the file, and I added the signature of functions
in the documentation of within(), get(), set(), publish(), subscribe().

Note:
Unlike what was stated previously, anonymous modules are created
by calling within() *without* any parameter, not *with* any parameter.