Add callback parameter to `get()` for one-time subscription
This pattern is very useful for dependencies between modules:
a function declaration should be declared only once,
and the requiring module must get the definition only once,
without fear of getting registered and triggered in duplicate
in case the function declaration was published more than once.

This should prevent errors in case an event is initially published
only once, and subscribers expect it to be published only once,
registering additional subscribers in the callback, but the event
is later published more than once, which would result in duplicate
registrations of the subscribers in the callback.